requirements.txt
django~=4.1.6
pytz~=2022.1
djangorestframework~=3.14.0
django-rest-passwordreset
psycopg2-binary
pyyaml~=5.4.1
python-dotenv~=0.19.2
ujson~=5.7.0
requests~=2.25.1
pip~=23.0.1
attrs~=21.2.0
DateTime~=5.0
wheel~=0.37.1
cryptography~=3.4.8
Pillow~=9.0.1
keyring~=23.5.0
setuptools~=59.6.0
asgiref~=3.6.0
sqlparse~=0.4.3
psycopg2~=2.9.2
colorama~=0.4.4
click~=8.0.3
certifi~=2020.6.20
six~=1.16.0
idna~=3.3
urllib3~=1.26.5
chardet~=4.0.0
future~=0.18.2